feat: authenticated routing!
