+++ lib/app/app.dart
@@ -1,15 +1,99 @@
 /// The app widget.
+///
+/// Including the app's initiation code.
 library pirate_code.app;
 
+import "dart:async";
+import "dart:developer";
+
+import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
+import "package:flutter_web_plugins/url_strategy.dart";
 
-import "bootstrap.dart";
-import "view/app.dart";
+import "../../l10n/l10n.dart";
+import "app_router.dart";
 
-export "app_router.dart";
-
-/// The app widget, with [Bootstrap]pin' capabilities.
-class App extends ConsumerWidget with AppView, Bootstrap {
+/// The app widget, with bootstrappin' capabilities.
+class App extends ConsumerWidget {
   /// Create a new instance of [App].
   App({super.key});
+
+  /// The app's container.
+  final container = ProviderContainer();
+
+  // Make sure you don't initiate your router inside of the build function.
+  late final AppRouter _appRouter;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    const flutterLocale = Locale("en", "US");
+    final theme = ColorScheme.fromSeed(seedColor: Colors.green);
+
+    return MaterialApp.router(
+      routerConfig: _appRouter.config(),
+      builder: (context, child) {
+        return _MainArea(child: child);
+      },
+      title: "PSD Wallet",
+      theme: ThemeData(
+        useMaterial3: true,
+        colorScheme: theme,
+        appBarTheme: AppBarTheme(color: theme.primary),
+      ),
+      locale: flutterLocale,
+      localizationsDelegates: AppLocalizations.localizationsDelegates,
+      supportedLocales: AppLocalizations.supportedLocales,
+    );
+  }
+
+  /// Bootstrap the app.
+  ///
+  /// This function will handle the following:
+  /// - [FlutterError.onError] will be set to log errors to the console.
+  /// - [usePathUrlStrategy] will be called to use path-style URLs.
+  /// - [runZonedGuarded] will be used to catch errors and log them to the console.
+  /// - [runApp] will be called to run the app.
+  /// - [ProviderScope] will be used to wrap the app.
+  Future<void> bootstrap() async {
+    FlutterError.onError = (details) {
+      log(details.exceptionAsString(), stackTrace: details.stack);
+    };
+
+    // Don't use hash style routes.
+    usePathUrlStrategy();
+
+    _appRouter = AppRouter(container: container);
+
+    await runZonedGuarded(
+      () async => runApp(
+        UncontrolledProviderScope(
+          container: container,
+          child: this,
+        ),
+      ),
+      (error, stackTrace) => log(error.toString(), stackTrace: stackTrace),
+    );
+  }
+}
+
+// The container for the current page, with its background color and subtle switching animation.
+class _MainArea extends ConsumerWidget {
+  const _MainArea({
+    required this.child,
+    // Temporary ignore, see <dart-lang/sdk#49025>.
+    // ignore: unused_element
+    super.key,
+  });
+
+  final Widget? child;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    return ColoredBox(
+      color: colorScheme.surfaceVariant,
+      child: child,
+    );
+  }
 }
