+++ lib/features/auth/data/auth_data.dart
@@ -1,12 +1,11 @@
 /// The auth feature's data.
 library pirate_code.features.auth.data;
 
-import "dart:developer";
-
 import "package:appwrite/appwrite.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
 import "../../../utils/data/api.dart";
+import "../domain/auth_domain.dart";
 
 part "auth_data.g.dart";
 
@@ -15,10 +14,10 @@ part "auth_data.g.dart";
 // ignore: one_member_abstracts
 abstract interface class AuthRepository {
   /// Authenticate the user.
-  Future<void> authenticate();
+  Future<PirateUser> authenticate();
 }
 
-/// The default implementation of [AuthRepository]
+/// The default implementation of [AuthRepository].
 class AppwriteAuthRepository implements AuthRepository {
   /// Create a new instance of [AppwriteAuthRepository].
   AppwriteAuthRepository(this.client);
@@ -27,16 +26,19 @@ class AppwriteAuthRepository implements AuthRepository {
   final Client client;
 
   @override
-  Future<void> authenticate() async {
-    final account = Account(client);
+  Future<PirateUser> authenticate() async {
+    final session = Account(client);
 
-    // Go to Google account login page
-    await account.createOAuth2Session(
+    // Go to the Google account login page.
+    await session.createOAuth2Session(
       provider: "google",
       success: "${Uri.base.origin}/auth.html",
       failure: "${Uri.base}",
     );
-    log("${account.get()}");
+    final account = await session.get();
+    return PirateUser(
+      name: account.name,
+    );
   }
 }
