+++ test/app/app_test.dart
@@ -6,6 +6,8 @@ import "package:pirate_code/app/app.dart";
 import "package:pirate_code/dart_define.gen.dart";
 import "package:pirate_code/utils/data/secrets.dart";
 
+import "../helpers/helpers.dart";
+
 void main() {
   group("App", () {
     testWidgets("renders a Material App", (tester) async {
@@ -20,6 +22,15 @@ void main() {
         ),
       );
       expect(find.byType(MaterialApp), findsOneWidget);
+    });
+  });
+  group("Bootstrappin' Tests!", () {
+    testWidgets("Test the boots...", (tester) async {
+      final tested = App();
+      await tested.bootstrap(); // No exception, we hope.
+      await tester.pumpApp(App());
+      expect(find.byType(MaterialApp), findsOneWidget);
+      expect(find.byType(ProviderScope), findsOneWidget);
     });
   });
 }
